refactor: add abstractions for data retrieval
Before, raw and formatted data were retrieved in terms of tables and
summaries. I reshaped this to retrieve either raw or formatted data.

The reason for this is that the spreadsheet will only ever use
formatted data. The raw data, on the other hand, will only go into
the database.