chore(DataExtractor): rectify type hint issues
A great deal of type errors were being returned in consequence of
incorrect type assignments with respect to various subroutines within
the DataExtractor module. I have fixed these by either removing some
type hints altogether or tailoring the pre-existing ones to be correct.
The sole reason for the removal of said type hints was due to the fact
that the data type can already be inferred from the very name of the
variables