fix: remove default worksheet on created workbook
Adding new functionality unfortunately is attended by certain costs.
Creating a new workbook necessitates that a default worksheet is created.
As a result of having to create a new workbook when the spreadsheet file
does not exist on the system, the default has to be removed such that
the budget worksheet is the only worksheet that is present when the
spreadsheet file is created on the file system.