refactor: improve data retrieval abstraction
This change displaces the TableFacade module with a DataFacade.
The problem that the previous TableFacade abstraction posed was that it
was doing far too little and may have very well been considered a
middleman. The new DataFacade module retrieves the data for both the
table and the summary whereas the previous TableFacade only retrieved
data for the Table